+++ tests/src/tests/mod.rs
@@ -8,6 +8,8 @@ pub mod multipart;
 pub mod nested_path;
 pub mod path;
 pub mod query;
+pub mod request_body_limit;
+pub mod request_time_limit;
 pub mod scheme;
 pub mod typed_header;